requirements.txt
tensorflow-gpu==2.3.1
tfa-nightly==0.12.0.dev20201027222106
mock==4.0.1
tqdm==4.42.1
numpy==1.18.1
miniimagenettools==0.2.2
pandas==1.0.1
tensorflow_hub==0.10.0
scikit-learn==0.22.1
#  Just for plots and graphs
matplotlib==3.1.3
seaborn==0.10.0
# for face recognition
# face_recognition
